[ias-verify] move tests to separate file
